Fix wizard state sync and extract focus trap
