+++ src/routes/admin/auth-providers/+page.svelte
@@ -8,6 +8,7 @@
 	import type { PageData } from './$types';
 	import * as Select from '$lib/components/ui/select';
 	import { getCsrfToken } from '$lib/utils/csrf';
+	import { modalFocusTrap } from '$lib/utils/focus-trap';
 	import { logger } from '$lib/utils/logger.js';
 
 	type ProviderType =
@@ -29,9 +30,6 @@
 	let success = $state('');
 	let loading = $state(false);
 	let roleMappingError = $state('');
-
-	const FOCUSABLE_SELECTOR =
-		'button:not([disabled]), [href], input:not([disabled]), select:not([disabled]), textarea:not([disabled]), [tabindex]:not([tabindex="-1"])';
 
 	$effect(() => {
 		const mapping = formData.roleMapping?.trim();
@@ -124,68 +122,6 @@
 		showEditModal = false;
 		showDeleteModal = false;
 		selectedProvider = null;
-	}
-
-	function getFocusableElements(container: HTMLElement): HTMLElement[] {
-		return Array.from(container.querySelectorAll<HTMLElement>(FOCUSABLE_SELECTOR)).filter(
-			(element) =>
-				!element.hasAttribute('disabled') &&
-				element.getAttribute('aria-hidden') !== 'true' &&
-				element.tabIndex !== -1
-		);
-	}
-
-	function modalFocusTrap(node: HTMLElement) {
-		let previousActiveElement: HTMLElement | null =
-			typeof document !== 'undefined' ? (document.activeElement as HTMLElement | null) : null;
-
-		const focusInitialElement = () => {
-			const labelledBy = node.getAttribute('aria-labelledby');
-			const labelledElement =
-				labelledBy && typeof document !== 'undefined'
-					? (document.getElementById(labelledBy) as HTMLElement | null)
-					: null;
-			const focusTarget = getFocusableElements(node)[0] ?? labelledElement ?? node;
-
-			if (labelledElement && !labelledElement.hasAttribute('tabindex')) {
-				labelledElement.tabIndex = -1;
-			}
-
-			focusTarget.focus();
-		};
-
-		const handleKeydown = (event: KeyboardEvent) => {
-			if (event.key !== 'Tab') return;
-
-			const focusables = getFocusableElements(node);
-			if (focusables.length === 0) {
-				event.preventDefault();
-				node.focus();
-				return;
-			}
-
-			const first = focusables[0];
-			const last = focusables[focusables.length - 1];
-
-			if (event.shiftKey && document.activeElement === first) {
-				event.preventDefault();
-				last.focus();
-			} else if (!event.shiftKey && document.activeElement === last) {
-				event.preventDefault();
-				first.focus();
-			}
-		};
-
-		node.addEventListener('keydown', handleKeydown);
-		setTimeout(focusInitialElement, 0);
-
-		return {
-			destroy() {
-				node.removeEventListener('keydown', handleKeydown);
-				previousActiveElement?.focus();
-				previousActiveElement = null;
-			}
-		};
 	}
 
 	function normalizeRoleMappingForSave(value: string) {
